+++ frontend/wailsjs/go/main/App.js
@@ -58,6 +58,10 @@ export function GetAllBackups() {
   return window['go']['main']['App']['GetAllBackups']();
 }
 
+export function GetAllPortStates() {
+  return window['go']['main']['App']['GetAllPortStates']();
+}
+
 export function GetAuditLogs(arg1) {
   return window['go']['main']['App']['GetAuditLogs'](arg1);
 }
@@ -106,6 +110,10 @@ export function GetPlaybooks() {
   return window['go']['main']['App']['GetPlaybooks']();
 }
 
+export function GetPortStates(arg1) {
+  return window['go']['main']['App']['GetPortStates'](arg1);
+}
+
 export function GetScheduledJobs() {
   return window['go']['main']['App']['GetScheduledJobs']();
 }
@@ -136,6 +144,10 @@ export function RunBackup(arg1) {
 
 export function RunPlaybook(arg1) {
   return window['go']['main']['App']['RunPlaybook'](arg1);
+}
+
+export function RunPortTrackerScan(arg1) {
+  return window['go']['main']['App']['RunPortTrackerScan'](arg1);
 }
 
 export function RunSSHCommands(arg1) {